feat(diet): daily target check

+++ lib/src/main/kotlin/com/rmp/lib/shared/modules/diet/DietWaterLogModel.kt
@@ -6,5 +6,5 @@ import com.rmp.lib.utils.korm.IdTable
 object DietWaterLogModel: IdTable("diet_water_log_model") {
     val createdAt = long("createdAt")
     val userId = reference("user_id", UserModel)
-    val volume = float("volume")
+    val volume = double("volume")
 }
